safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/url-no-trailing-slash.html

<!DOCTYPE">
<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("URL that doesn't have trailing slash should not emit empty Request-URI.");

window.jsTestIsAsync = true;

var url = "ws://127.0.0.1:8880";
var handshake_success = false;
var host

var ws = new WebSocket(url);
ws.onopen = function () {
    debug("WebSocket is open");
    handshake_success = true;
};
ws.onmessage = function (evt) {
    host = evt.data;
    debug("received:" + host);
    ws.close();
};
ws.onclose = function () {
    debug("WebSocket is closed");
    shouldBeTrue("handshake_success");
    shouldBe("host", '"127.0.0.1:8880"');
    finishJSTest();
};

</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
